output: use float instead of half

+++ include/openposert/output/nms.hpp
@@ -2,13 +2,11 @@
 
 #include <array>
 
-#include "cuda_fp16.h"
-
 namespace openposert {
 
-void nms(__half* target_ptr, int* kernel_ptr, const __half* const source_ptr,
-         const __half threshold, const std::array<int, 4>& target_size,
-         const std::array<int, 4>& source_size, const __half offset_x,
-         const __half offset_y);
+void nms(float* target_ptr, int* kernel_ptr, const float* const source_ptr,
+         const float threshold, const std::array<int, 4>& target_size,
+         const std::array<int, 4>& source_size, const float offset_x,
+         const float offset_y);
 
 }  // namespace openposert
